Added reduced linenoise implementation for stdin/stdout command line input. Example:
	var linenoise = require('linenoise');
	linenoise.setPrompt('test> ');
	setInterval(function() {
		var line = linenoise.poll();
		if(line!==undefined) linenoise.print(line);
	}, 20);

We need to poll it every so often so it can process key presses. The poll
function will return a complete line when one is available.

module_aes.c had some leftover includes that it didn't need.

Had to modify the partitions.csv to use all the available space. The
linenoise addition puts us over the 1M size.